frontend upgrades, email, author

+++ admin/includes/view_all_users.php
@@ -63,11 +63,17 @@
 
         if(isset($_GET["delete"]))
         {
-            $user_id = $_GET["delete"];
-            $query = "DELETE FROM users WHERE user_id = {$user_id}";
-            $delete_query = mysqli_query($connection, $query);
-            confirm_query($delete_query);
-            header("Location: users.php");    
+            if(isset($_SESSION["user_role"]))
+            {
+                if($_SESSION["user_role"] == "admin")
+                {
+                    $user_id = mysqli_real_escape_string($connection, $_GET["delete"]);
+                    $query = "DELETE FROM users WHERE user_id = {$user_id}";
+                    $delete_query = mysqli_query($connection, $query);
+                    confirm_query($delete_query);
+                    header("Location: users.php");
+                }
+            } 
         }
 
         if(isset($_GET["change_to_admin"]))
